backend/requirements.txt
flask==3.1.0
flask-cors==5.0.1
python-dotenv==1.0.1
requests==2.32.3
sentence-transformers==3.3.1
faiss-cpu==1.9.0.post1
pandas==2.2.3
PyPDF2==3.0.1
python-docx==1.1.2
openpyxl==3.1.5
reportlab==4.2.5
gunicorn==23.0.0
